Changed Class names and Small Refactor
